Nira Discord bot/Nira-bot-main/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
astunparse==1.6.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
braceexpand==0.1.7
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
decorator==5.1.1
defusedxml==0.7.1
discord-components==0.0.0.1
discord.py==2.4.0
docopt==0.6.2
emoji==2.12.1
executing==2.0.1
fastjsonschema==2.20.0
Flask==3.0.3
frozenlist==1.4.1
idna==3.7
import_expression==1.1.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jishaku==2.5.2
joblib==1.4.2
jokeapi==1.0.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.0.1
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycountry==24.6.1
pycparser==2.22
Pygments==2.18.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pyttsx3==2.90
pytz==2024.1
pyzmq==26.1.0
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
scikit-learn==1.5.1
scipy==1.14.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
spotipy==2.24.0
stack-data==0.6.3
threadpoolctl==3.5.0
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==1.26.19
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
Werkzeug==3.0.3
wheel==0.43.0
yarg==0.1.9
yarl==1.9.4
qrcode


